+++ apis/userApis.js
@@ -1,6 +1,8 @@
 var Users = require('../models/users.js');
 var Employers = require('../models/employers.js');
 
+var bcrypt = require('bcryptjs');
+
 
 // get helper functions
 const { sanitize,
@@ -21,7 +23,8 @@ const userApis = {
     GET_keepMeLoggedIn,
     GET_session,
     POST_session,
-    POST_verifyEmail
+    POST_verifyEmail,
+    POST_changePasswordForgot
 }
 
 
@@ -188,6 +191,64 @@ function POST_verifyEmail(req, res) {
                 return res.json("go to login");
             }
         });
+    });
+}
+
+
+function POST_changePasswordForgot(req, res) {
+    let token = sanitize(req.body.token).toString();
+    let password = sanitize(req.body.password);
+
+    var query = {passwordToken: token};
+    Users.findOne(query, function (err, user) {
+        if (err) {
+            console.log("Error trying to find user from password token: ", err);
+            return res.status(500).send("Server error, try again later");
+        }
+
+        if (!user) {
+            return res.status(404).send("User not found from link");
+        }
+
+        const currentTime = Date.now();
+        if (currentTime > user.passwordTokenExpirationTime) {
+            return res.status(401).send("Time ran out, try sending email again");
+        }
+
+        let query = {_id: user._id};
+        const saltRounds = 10;
+        bcrypt.genSalt(saltRounds, function (err, salt) {
+            bcrypt.hash(password, salt, function (err, hash) {
+                // change the stored password to be the hash
+                const newPassword = hash;
+                // if the field doesn't exist, $set will set a new field
+                // can be verified because the user had to go to their email
+                // to get to this page
+                var update = {
+                    '$set': {
+                        password: newPassword,
+                        verified: true
+                    },
+                    '$unset': {
+                        passwordToken: "",
+                        passwordTokenExpirationTime: "",
+                    }
+                };
+
+                // When true returns the updated document
+                var options = {new: true};
+
+                Users.findOneAndUpdate(query, update, options, function (err, newUser) {
+                    if (err) {
+                        console.log(err);
+                        return res.status(500).send("Error saving new password");
+                    }
+
+                    // successfully created new password
+                    return res.json(removePassword(newUser));
+                });
+            })
+        })
     });
 }
