+++ src/components/pages/changepasswordforgot.js
@@ -3,7 +3,7 @@ import React, {Component} from 'react';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import {changePasswordForgot} from '../../actions/usersActions';
-import {TextField} from 'material-ui';
+import {TextField, CircularProgress} from 'material-ui';
 import {Field, reduxForm} from 'redux-form';
 import HomepageTriangles from '../miscComponents/HomepageTriangles';
 import MetaTags from 'react-meta-tags';
@@ -109,6 +109,11 @@ class PasswordChange extends Component {
                         >
                             Change Password
                         </button>
+                        <br/>
+                        {this.props.loadingChangePassword ?
+                            <CircularProgress />
+                            : null
+                        }
                     </form>
                 </div>
             </div>
@@ -125,7 +130,8 @@ function mapDispatchToProps(dispatch) {
 function mapStateToProps(state) {
     return {
         formData: state.form,
-        currentUser: state.users.currentUser
+        currentUser: state.users.currentUser,
+        loadingChangePassword: state.users.loadingSomething
     };
 }
